# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Music recorder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Song {
  id: number;
  title: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  tags?: string | null;
  /**
     * Duration in seconds
     * @nullable
     */
  duration?: number | null;
  /**
     * URL to stream the audio file
     * @nullable
     */
  audioUrl?: string | null;
  hasAudio: boolean;
  /**
     * Amplitude peak values (0–1) for waveform visualization
     * @nullable
     */
  waveformData?: number[] | null;
  /** Whether this song is listed on the open sessions board */
  isPublic: boolean;
  /**
     * What kind of collaboration the artist is looking for
     * @nullable
     */
  seekingHelp?: string | null;
  /**
     * Share token for collaboration link
     * @nullable
     */
  shareToken?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface SongInput {
  /** @minLength 1 */
  title: string;
  notes?: string;
  tags?: string;
}

export interface SongUpdate {
  /** @minLength 1 */
  title?: string;
  notes?: string;
  tags?: string;
}

export interface PublishInput {
  isPublic: boolean;
  /** What kind of help the artist is looking for (e.g. "lead guitar", "vocals") */
  seekingHelp?: string;
}

export interface ShareResult {
  shareToken: string;
  shareUrl: string;
}

export interface Session {
  id: number;
  title: string;
  hasAudio: boolean;
  isPublic: boolean;
  /** @nullable */
  seekingHelp?: string | null;
  /** @nullable */
  duration?: number | null;
  /** Number of collaboration tracks submitted */
  collabCount: number;
  shareToken: string;
  /** @nullable */
  audioUrl?: string | null;
  createdAt: string;
}

export interface CollabSong {
  id: number;
  title: string;
  /** @nullable */
  seekingHelp?: string | null;
  hasAudio: boolean;
  /** @nullable */
  duration: number | null;
  /** @nullable */
  audioUrl: string | null;
}

export interface CollabTrack {
  id: number;
  /** @nullable */
  authorName?: string | null;
  audioUrl: string;
  /** @nullable */
  duration?: number | null;
  createdAt: string;
}

export interface SongStats {
  totalSongs: number;
  /** Total duration of all songs in seconds */
  totalDuration: number;
  songsWithAudio: number;
  recentSongs: Song[];
}