# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Music recorder API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: songs
    description: Song management
  - name: sessions
    description: Public open sessions board
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /songs:
    get:
      operationId: listSongs
      tags: [songs]
      summary: List all songs
      responses:
        "200":
          description: List of songs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Song"
    post:
      operationId: createSong
      tags: [songs]
      summary: Create a new song
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SongInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"

  /songs/stats:
    get:
      operationId: getSongStats
      tags: [songs]
      summary: Get song statistics
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SongStats"

  /songs/{id}/share:
    post:
      operationId: shareSong
      tags: [songs]
      summary: Generate or retrieve a share token for a song
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Share token and URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareResult"
        "404":
          description: Not found

  /songs/{id}/publish:
    post:
      operationId: publishSong
      tags: [songs]
      summary: Toggle a song's public visibility on the open sessions board
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishInput"
      responses:
        "200":
          description: Updated song
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
        "404":
          description: Not found

  /sessions:
    get:
      operationId: listSessions
      tags: [sessions]
      summary: List all public open sessions
      responses:
        "200":
          description: Public songs open for collaboration
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"

  /collab/{token}:
    get:
      operationId: getCollabSong
      tags: [songs]
      summary: Get song info by share token (public)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Song info for collaboration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollabSong"
        "404":
          description: Not found

  /collab/{token}/tracks:
    get:
      operationId: listCollabTracks
      tags: [songs]
      summary: List collaboration tracks for a shared song
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of collaboration tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CollabTrack"

  /songs/{id}:
    get:
      operationId: getSong
      tags: [songs]
      summary: Get a single song
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Song
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
        "404":
          description: Not found
    patch:
      operationId: updateSong
      tags: [songs]
      summary: Update a song
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SongUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
        "404":
          description: Not found
    delete:
      operationId: deleteSong
      tags: [songs]
      summary: Delete a song
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Song:
      type: object
      required: [id, title, hasAudio, isPublic, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        notes:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        duration:
          type: ["number", "null"]
          description: Duration in seconds
        audioUrl:
          type: ["string", "null"]
          description: URL to stream the audio file
        hasAudio:
          type: boolean
        waveformData:
          type: ["array", "null"]
          items:
            type: number
          description: Amplitude peak values (0–1) for waveform visualization
        isPublic:
          type: boolean
          description: Whether this song is listed on the open sessions board
        seekingHelp:
          type: ["string", "null"]
          description: What kind of collaboration the artist is looking for
        shareToken:
          type: ["string", "null"]
          description: Share token for collaboration link
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    SongInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        notes:
          type: string
        tags:
          type: string

    SongUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        notes:
          type: string
        tags:
          type: string

    PublishInput:
      type: object
      required: [isPublic]
      properties:
        isPublic:
          type: boolean
        seekingHelp:
          type: string
          description: What kind of help the artist is looking for (e.g. "lead guitar", "vocals")

    ShareResult:
      type: object
      required: [shareToken, shareUrl]
      properties:
        shareToken:
          type: string
        shareUrl:
          type: string

    Session:
      type: object
      required: [id, title, hasAudio, isPublic, collabCount, shareToken, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        hasAudio:
          type: boolean
        isPublic:
          type: boolean
        seekingHelp:
          type: ["string", "null"]
        duration:
          type: ["number", "null"]
        collabCount:
          type: integer
          description: Number of collaboration tracks submitted
        shareToken:
          type: string
        audioUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    CollabSong:
      type: object
      required: [id, title, hasAudio, duration, audioUrl]
      properties:
        id:
          type: integer
        title:
          type: string
        seekingHelp:
          type: ["string", "null"]
        hasAudio:
          type: boolean
        duration:
          type: ["number", "null"]
        audioUrl:
          type: ["string", "null"]

    CollabTrack:
      type: object
      required: [id, audioUrl, createdAt]
      properties:
        id:
          type: integer
        authorName:
          type: ["string", "null"]
        audioUrl:
          type: string
        duration:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time

    SongStats:
      type: object
      required: [totalSongs, totalDuration, songsWithAudio, recentSongs]
      properties:
        totalSongs:
          type: integer
        totalDuration:
          type: number
          description: Total duration of all songs in seconds
        songsWithAudio:
          type: integer
        recentSongs:
          type: array
          items:
            $ref: "#/components/schemas/Song"