# lib/api-zod/src/generated/types/publishInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Music recorder API
 * OpenAPI spec version: 0.1.0
 */

export interface PublishInput {
  isPublic: boolean;
  /** What kind of help the artist is looking for (e.g. "lead guitar", "vocals") */
  seekingHelp?: string;
}