# lib/api-zod/src/generated/types/session.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Music recorder API
 * OpenAPI spec version: 0.1.0
 */

export interface Session {
  id: number;
  title: string;
  hasAudio: boolean;
  isPublic: boolean;
  /** @nullable */
  seekingHelp?: string | null;
  /** @nullable */
  duration?: number | null;
  /** Number of collaboration tracks submitted */
  collabCount: number;
  shareToken: string;
  /** @nullable */
  audioUrl?: string | null;
  createdAt: Date;
}