# lib/api-zod/src/generated/types/song.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Music recorder API
 * OpenAPI spec version: 0.1.0
 */

export interface Song {
  id: number;
  title: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  tags?: string | null;
  /**
     * Duration in seconds
     * @nullable
     */
  duration?: number | null;
  /**
     * URL to stream the audio file
     * @nullable
     */
  audioUrl?: string | null;
  hasAudio: boolean;
  /**
     * Amplitude peak values (0–1) for waveform visualization
     * @nullable
     */
  waveformData?: number[] | null;
  /** Whether this song is listed on the open sessions board */
  isPublic: boolean;
  /**
     * What kind of collaboration the artist is looking for
     * @nullable
     */
  seekingHelp?: string | null;
  /**
     * Share token for collaboration link
     * @nullable
     */
  shareToken?: string | null;
  createdAt: Date;
  updatedAt: Date;
}